feat: Android implementation for dynamic session controls customization
