+++ android/src/main/java/io/cobrowse/reactnative/CobrowseIODelegatesWithSessionControl.java
@@ -1,0 +1,6 @@
+package io.cobrowse.reactnative;
+
+import io.cobrowse.CobrowseIO;
+
+public interface CobrowseIODelegatesWithSessionControl extends CobrowseIOCommonDelegates, CobrowseIO.SessionControlsDelegate {
+}
